--- a/logstash-core/src/main/java/org/logstash/PathCache.java
+++ b/logstash-core/src/main/java/org/logstash/PathCache.java
@@ -5,34 +5,24 @@ import java.util.concurrent.ConcurrentHashMap;
 
 public final class PathCache {
 
-    private static final Map<String, FieldReference> cache =
+    private static final Map<CharSequence, FieldReference> CACHE =
         new ConcurrentHashMap<>(64, 0.2F, 1);
 
-    private static final FieldReference timestamp = cache(Event.TIMESTAMP);
-
-    private static final String BRACKETS_TIMESTAMP = "[" + Event.TIMESTAMP + "]";
-
-    static {
-        // inject @timestamp
-        cache.put(BRACKETS_TIMESTAMP, timestamp);
+    private PathCache() {
     }
 
-    public static boolean isTimestamp(String reference) {
-        return cache(reference) == timestamp;
-    }
-
-    public static FieldReference cache(String reference) {
+    public static FieldReference cache(final CharSequence reference) {
         // atomicity between the get and put is not important
-        final FieldReference result = cache.get(reference);
+        final FieldReference result = CACHE.get(reference);
         if (result != null) {
             return result;
         }
         return parseToCache(reference);
     }
-    
-    private static FieldReference parseToCache(final String reference) {
+
+    private static FieldReference parseToCache(final CharSequence reference) {
         final FieldReference result = FieldReference.parse(reference);
-        cache.put(reference, result);
+        CACHE.put(reference, result);
         return result;
     }
 }