… logstash-core/src/test/java/org/logstash/FieldReferenceTest.java