This change provides for a way more efficient lookup strategy than the current implementation or even #7847 by:

* Lookup by `FieldReference` and use the fact that the `FieldReference` already contains knowledge about a few things (metadata, timestamp, tag field and so on)
* Deduplicate `FieldReference` globally to get better cache behavior

This avoids:

* A `RubyString` to `String` conversion on **every** `Event` field access!
* Needless string comparisons to handle the metadata corner cases
* Redundant lookup of the `FieldReference` in `RubyEvent#set` (we did look it up once for the timestamp check and then once more for the actual `Event#set` operation downstream)
* Having a lot of logic for `String` parsing that messes with inlining in the hot path (note that we already have 100% of the parse results cached when we do the `String` operations on `reference` in the `Event` lookups/sets at the moment and are fixing this here)

Also, this allows the `RubyString` key to be GCed much sooner if it's not a reused value (which I suspect is the case for many many use cases).

### Before

```
bin/logstash -b 128 -w 4 -f ~/tmp/logstash.cfg  1706.14s user 36.36s system 417% cpu 6:57.64 total
pv  4.32s user 26.83s system 7% cpu 6:57.64 total
wc -c  0.61s user 5.41s system 1% cpu 6:57.64 total
```

### After

```
➜  logstash git:(lut-no-more2-rebase) ✗ time cat ~/Downloads/apache_access_logs ~/Downloads/apache_access_logs | bin/logstash -b 128 -w 4 -f ~/tmp/logstash.cfg  | pv  | wc -c
bin/logstash -b 128 -w 4 -f ~/tmp/logstash.cfg  1620.05s user 32.34s system 412% cpu 6:40.48 total
pv  3.80s user 24.19s system 6% cpu 6:40.48 total
wc -c  0.53s user 4.35s system 1% cpu 6:40.48 total
```

Note that this includes 30s ~ compilation/warmup => we're 6:10 vs. 6:27, which isn't so shabby considering that we're also putting less load on the CPU from fewer GCs.
Stats here

* Collections (only minor happening) 351/min down to 332/min
* Average pause stays the same at ~10ms
* Number of overall pauses in the run goes down from 2.2k to 2k (10% reduction!)

This should be (haven't tried it yet, but even the much less efficient #7847 showed a neat aggregate improvement) a lot more fun combined with #7854