@guyboertje 2017-08-03
I like this!!
its all very sensible.
FieldReference dedup ++
Answers the questions:
- What is a FieldReference really?
- What is the PathCache really?