@guyboertje 2017-08-03
P.S. I think another review is very optional. This is all internal, no Event API changes for Java plugin authors.