@suyograo 2017-08-03
@original-brownbear backport to 6.0 and 5.6 as well please.