@original-brownbear 2017-08-03
@suyograo backporting this to `5.6` entails backporting at least also backporting #7839, still do it?