@original-brownbear 2017-08-04
done => links:

5.6: #7909 
6.0: #7910